+++ public/includes/class-wp-ng-public-fallback.php
@@ -39,21 +39,29 @@ class Wp_Ng_Public_Fallback {
 
     wp_deregister_script('angular');
 
-    $bower = new Wp_Ng_Bower();
-
-    wp_register_script(
-      'angular',
-      $bower->map_to_cdn([
+    if ( wp_ng_is_cdn_angular() ) {
+      $bower = new Wp_Ng_Bower();
+      $src = $bower->map_to_cdn([
         'name' => 'angular',
         'cdn' => 'google-angular',
         'file' => 'angular.min.js'
-      ], wp_ng_get_asset_path('scripts/angular.js')),
+      ], wp_ng_get_asset_path('scripts/angular.js'));
+    }
+    else {
+      $src = wp_ng_get_asset_path('scripts/angular.js');
+    }
+
+    wp_register_script(
+      'angular',
+      $src,
       $deps,
       $bower->get_version('angular'),
       true
     );
 
-    wp_add_inline_script( 'angular', '!window.angular && document.write("\x3Cscript src=\x22' . wp_ng_get_asset_path('scripts/angular.js') . '\x22\x3E\x3C/script\x3E");' );
+    if ( wp_ng_is_cdn_angular() ) {
+      wp_add_inline_script( 'angular', '!window.angular && document.write("\x3Cscript src=\x22' . wp_ng_get_asset_path('scripts/angular.js') . '\x22\x3E\x3C/script\x3E");' );
+    }
   }
 
   /**
@@ -63,28 +71,75 @@ class Wp_Ng_Public_Fallback {
    */
   static public function register_jquery_fallback( $deps = array()) {
 
-    if( !is_array($deps) ) {
-      $deps = array(strval($deps));
+    global $wp_scripts;
+
+    if ( ! is_array( $deps ) ) {
+      $deps = array( strval( $deps ) );
     }
 
-    wp_deregister_script('jquery');
+    $bower   = new Wp_Ng_Bower();
+    $version = $bower->get_version( 'jquery' );
 
-    $bower = new Wp_Ng_Bower();
-    $version = $bower->get_version('jquery');
+    wp_deregister_script( 'jquery' );
+    wp_deregister_script( 'jquery-core' );
 
-    wp_register_script(
-      'jquery',
-      $bower->map_to_cdn([
+    if ( wp_ng_is_cdn_jquery() ) {
+
+      $src = $bower->map_to_cdn( [
         'name' => 'jquery',
-        'cdn' => 'jquery',
+        'cdn'  => 'jquery',
         'file' => 'jquery-' . $version . '.min.js'
-      ], wp_ng_get_asset_path('scripts/jquery.js')),
+      ], wp_ng_get_asset_path( 'scripts/jquery.js' ) );
+    }
+    else {
+      $src = wp_ng_get_asset_path( 'scripts/jquery.js' );
+    }
+
+    wp_register_script( 'jquery', false, array( 'jquery-core', 'jquery-migrate' ), $version );
+    wp_register_script(
+      'jquery-core',
+      $src,
       $deps,
       $version,
       true
     );
 
-    wp_add_inline_script( 'jquery', '!window.jQuery && document.write("\x3Cscript src=\x22' . wp_ng_get_asset_path('scripts/jquery.js') . '\x22\x3E\x3C/script\x3E");' );
+    if ( wp_ng_is_cdn_jquery() ) {
+      wp_add_inline_script( 'jquery-core', '!window.jQuery && document.write("\x3Cscript src=\x22' . wp_ng_get_asset_path( 'scripts/jquery.js' ) . '\x22\x3E\x3C/script\x3E");' );
+    }
+
+    //Add jquery migrate to cdn with fallback
+    if ( isset($wp_scripts->registered['jquery-migrate']) ) {
+
+      $original_src = $wp_scripts->registered['jquery-migrate']->src;
+      $version = $wp_scripts->registered['jquery-migrate']->ver;
+
+      if ( wp_ng_is_cdn_jquery() ) {
+        $src = $bower->map_to_cdn( [
+          'name' => 'jquery-migrate',
+          'cdn'  => 'jquery-migrate',
+          'file' => 'jquery-migrate-' . $version . '.min.js'
+        ], $original_src, $version );
+      }
+      else {
+        $src = $original_src;
+      }
+
+      wp_deregister_script( 'jquery-migrate' );
+      wp_register_script(
+        'jquery-migrate',
+        $src,
+        array(),
+        $version,
+        true
+      );
+
+      if ( wp_ng_is_cdn_jquery() ) {
+        wp_add_inline_script( 'jquery-migrate', '!window.jQuery.migrateWarnings  && document.write("\x3Cscript src=\x22' . get_home_url(null, $src) . '\x22\x3E\x3C/script\x3E");' );
+      }
+
+    }
+
   }
 
 }